requirements.txt
altgraph==0.17.2
appnope==0.1.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
macholib==1.16
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
requests==2.28.0
six==1.16.0
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
xerox==0.4.1
